Another version of the events stuff
It directly calls the methods in the modules
Also made a fun module
