+++ src/main/java/buttondevteam/discordplugin/listeners/DiscordListener.java
@@ -1,0 +1,4 @@
+package buttondevteam.discordplugin.listeners;
+
+public interface DiscordListener {
+}
