+++ src/main/java/buttondevteam/discordplugin/mcchat/MinecraftChatModule.java
@@ -3,13 +3,19 @@ package buttondevteam.discordplugin.mcchat;
 import buttondevteam.discordplugin.DiscordPlugin;
 import buttondevteam.lib.TBMCCoreAPI;
 import buttondevteam.lib.architecture.Component;
+import lombok.Getter;
 
 public class MinecraftChatModule extends Component {
+	private @Getter MCChatListener listener;
+
+	public MCChatListener getListener() { //It doesn't want to generate
+		return listener;
+	}
 	@Override
 	protected void enable() {
-		MCChatListener mcchat = new MCChatListener();
-		DiscordPlugin.dc.getDispatcher().registerListener(mcchat);
-		TBMCCoreAPI.RegisterEventsForExceptions(mcchat, getPlugin());
+		listener = new MCChatListener();
+		DiscordPlugin.dc.getDispatcher().registerListener(listener);
+		TBMCCoreAPI.RegisterEventsForExceptions(listener, getPlugin());
 		TBMCCoreAPI.RegisterEventsForExceptions(new MCListener(), getPlugin());
 	}
